move Field datatype to Internal.hs
necessary for next commit
